Improve perf of output semantic tokens
