--- src/Compiler/CompiledFileOutputNonSerializedMetadata.cs
@@ -1,8 +1,0 @@
-﻿using Microsoft.CodeAnalysis.Classification;
-
-namespace DotNetLab;
-
-internal sealed class CompiledFileOutputNonSerializedMetadata
-{
-    public ImmutableArray<ClassifiedSpan> ClassifiedSpans { get; init; }
-}
