+++ src/WorkerApi/InputMessage.cs
@@ -18,7 +18,6 @@ namespace DotNetLab;
 [JsonDerivedType(typeof(ProvideCompletionItems), nameof(ProvideCompletionItems))]
 [JsonDerivedType(typeof(ResolveCompletionItem), nameof(ResolveCompletionItem))]
 [JsonDerivedType(typeof(ProvideSemanticTokens), nameof(ProvideSemanticTokens))]
-[JsonDerivedType(typeof(ProvideOutputSemanticTokens), nameof(ProvideOutputSemanticTokens))]
 [JsonDerivedType(typeof(ProvideCodeActions), nameof(ProvideCodeActions))]
 [JsonDerivedType(typeof(ProvideHover), nameof(ProvideHover))]
 [JsonDerivedType(typeof(ProvideSignatureHelp), nameof(ProvideSignatureHelp))]
@@ -147,14 +146,6 @@ public abstract record WorkerInputMessage
         }
     }
 
-    public sealed record ProvideOutputSemanticTokens(string ModelUri, bool Debug) : WorkerInputMessage<string?>
-    {
-        public override Task<string?> HandleAsync(IExecutor executor)
-        {
-            return executor.HandleAsync(this);
-        }
-    }
-
     public sealed record ProvideCodeActions(string ModelUri, string? RangeJson) : WorkerInputMessage<string?>
     {
         public override Task<string?> HandleAsync(IExecutor executor)
@@ -217,7 +208,6 @@ public abstract record WorkerInputMessage
         Task<string> HandleAsync(ProvideCompletionItems message);
         Task<string?> HandleAsync(ResolveCompletionItem message);
         Task<string?> HandleAsync(ProvideSemanticTokens message);
-        Task<string?> HandleAsync(ProvideOutputSemanticTokens message);
         Task<string?> HandleAsync(ProvideCodeActions message);
         Task<string?> HandleAsync(ProvideHover message);
         Task<string?> HandleAsync(ProvideSignatureHelp message);
